1. Use code style to format code. 2. Simplify package and module name.
